Abstract double buffering for MPI
Implements a Buffer type that can be used with MPI
to implement efficient send and receives.

If the user not using GPUs or has CUDA-aware MPI
implementation they can use `BufferKind::Single`
to avoid the extra copy required by double buffering.

Using `BufferKind::Double` will pin the host side
of the buffer to get faster copies between the host
and the device.

For now we default to double buffering for the CPU
and single buffering for CPU only.

Co-authored-by: Kiran Pamnany <kpamnany@gmail.com>