Use asynchronous memory transfer for Buffer
The overall design is that memory copies should
be asynchronous with regard to non-blocking streams.
This enables us to run compute kernels in parallel to
memory transfers. Since CUDA-aware MPI will use
`CuDefaultStream` and has thus to be avoided for compute
kernels in the first place, we place these memory transfer
onto `CuDefaultStream` as well.

The buffers are used by MPI and we need to ensure that
the memory copies are done before we issue the `ISend`s.
Intuitively that would mean we need to synchronize the stream.
We have observed the the use of `synchronize(CuDefaultStream())`,
can lead to deadlocks (maybe just on our over-subscribed CI system),
and so we have to use a `query(stream); IProbe()` loop to enable
progress in MPI.

The receiving side is synchronized right now through the default stream,
since the sequence of events is

```
finish_ghost_exchange
  Waitall!
  prepare_stage # async copy on CuDefaultStream
  transferrecvbuf! # kernel launch on CuDefaultStream

faceviscterms # compute kernel launch on CuDefaultStream

start_ghost_exchange
  fillsendbuf! # kernel launch on CuDefaultStream
  prepare_transfer # async_copy on CuDefaultStream
    mpi_friendly_synchronize # synchronize CuDefaultStream
  ISend!
```

Since `faceviscterms` has a dependency on `transferrecvbuf!`,
and `fillsendbug!` has a dependency on `faceviscterms`, we either
need to place them on the same stream (as they are currently),
or add dependency edges through events.